requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-redis
psycopg2-binary
redis
dj-database-url
